+++ Setup.hs
@@ -1,4 +1,9 @@
 import Distribution.Simple
 import Language.Java.Inline.Cabal
+import qualified Language.Java.Streaming.Jars (getJars)
 
-main = defaultMainWithHooks (gradleHooks simpleUserHooks)
+main = do
+    jars <- Language.Java.Streaming.Jars.getJars
+    defaultMainWithHooks
+      $ addJarsToClasspath jars
+      $ gradleHooks simpleUserHooks